refactor: extract cohesion into analyzer/cohesion.rs (#190)
Moves CohesionMetrics + compute_cohesion out of analyzer/mod.rs.
audit() now calls the helper instead of inlining the loop.
